requirements.txt
streamlit>=1.30.0
openai>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
scikit-learn>=1.3.0
thefuzz[speedup]>=0.20.0
pytest>=7.4.0
tiktoken>=0.5.0
